Support newer libnx version
It appears that `swkbdConfigSetStringLenMaxExt` was replaced with `swkbdConfigSetStringLenMin` on the libnx repo... I was having trouble building the demo because of this. I also noticed std::optional was used but not included in the dx_renderer.hpp file? Not sure if thats a change that should be kept, but that was also causing problems for me.